♻️  ルーティングとプロバイダーの構成をリファクタリング
